+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Removed
 
 ### Fixed
+- Return ErrorResult from uncaught Action exceptions
 
 ### Tech-Debt/Refactor
 - Redis connection pool scales to action count, and no longer used by ingress